+++ outputs.tf
@@ -16,11 +16,11 @@ output "stac_server_api_domain_name" {
   value = (
     local.is_private_endpoint
     ? replace(
-      element(split("/", aws_api_gateway_deployment.stac_server_api_gateway.invoke_url), 2),
+      element(split("/", aws_api_gateway_stage.stac_server_api_gateway_stage.invoke_url), 2),
       aws_api_gateway_rest_api.stac_server_api_gateway.id,
-      "${aws_api_gateway_rest_api.stac_server_api_gateway.id}-${aws_vpc_endpoint.stac_server_api_gateway_private[0].id}"
+      "${aws_api_gateway_rest_api.stac_server_api_gateway.id}-${local.vpce_id}"
     )
-    : element(split("/", aws_api_gateway_deployment.stac_server_api_gateway.invoke_url), 2)
+    : element(split("/", aws_api_gateway_stage.stac_server_api_gateway_stage.invoke_url), 2)
   )
 }
 
@@ -28,7 +28,7 @@ output "stac_server_api_path" {
   value = "/${element(
     split(
       "/",
-      aws_api_gateway_deployment.stac_server_api_gateway.invoke_url,
+      aws_api_gateway_stage.stac_server_api_gateway_stage.invoke_url,
     ),
     3,
   )}"