+++ utils/cicd/main.tf
@@ -70,12 +70,16 @@ module "main" {
   authorized_s3_arns                        = []
   private_certificate_arn                   = ""
   vpce_private_dns_enabled                  = false
+  custom_vpce_id                            = null
   domain_alias                              = ""
   stac_api_url                              = ""
   # stac_docs_url
   # stac_api_stage_description
-  cors_origin      = "*"
-  cors_credentials = false
-  cors_methods     = ""
-  cors_headers     = ""
+  cors_origin               = "*"
+  cors_credentials          = false
+  cors_methods              = ""
+  cors_headers              = ""
+  asset_proxy_bucket_option = "NONE"
+  asset_proxy_bucket_list   = ""
+  asset_proxy_url_expiry    = 300
 }